Move atelet code to work with object storage under atelet/internal.
Also add otel traces when uploading objects, not just when downloading. This is useful when looking at traces during suspend.
