+++ cmd/atelet/internal/ategcs/gcs.go
@@ -1,0 +1,49 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package ategcs
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"io"
+
+	"cloud.google.com/go/storage"
+)
+
+type gcsClient struct {
+	client *storage.Client
+}
+
+func NewGCSClient(client *storage.Client) ObjectStorage {
+	return &gcsClient{client: client}
+}
+
+func (g *gcsClient) GetObject(ctx context.Context, bucket, object string) (io.ReadCloser, error) {
+	return g.client.Bucket(bucket).Object(object).NewReader(ctx)
+}
+
+func (g *gcsClient) PutObject(ctx context.Context, bucket, object string, reader io.Reader) error {
+	wc := g.client.Bucket(bucket).Object(object).NewWriter(ctx)
+	// io.Copy reports local read errors; wc.Close() reports the actual
+	// GCS upload (auth, permissions, transient). Join both so the caller
+	// doesn't lose either.
+	_, copyErr := io.Copy(wc, reader)
+	closeErr := wc.Close()
+	if err := errors.Join(copyErr, closeErr); err != nil {
+		return fmt.Errorf("while putting GCS object: %w", err)
+	}
+	return nil
+}
